cc: Enable chromium_code:1 for cc.
Most of the build fixes have landed without the flag turned on for
now. This has the remaining fixes to get this flag enabled.

R=jamesr
BUG=


Review URL: https://chromiumcodereview.appspot.com/13297005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191546 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d09344fde8e7822d9b91a571414f7ef49a36d973

+++ cc/output/geometry_binding.cc
@@ -26,7 +26,7 @@ GeometryBinding::GeometryBinding(WebKit::WebGraphicsContext3D* context,
     Vertex v0, v1, v2, v3;
   };
   struct QuadIndex {
-    uint16_t data[6];
+    uint16 data[6];
   };
 
   COMPILE_ASSERT(
@@ -55,12 +55,12 @@ GeometryBinding::GeometryBinding(WebKit::WebGraphicsContext3D* context,
                   i * 4.0f + 3.0f };
     Quad x = { v0, v1, v2, v3 };
     quad_list[i] = x;
-    QuadIndex y = { { 0 + 4 * i,
-                      1 + 4 * i,
-                      2 + 4 * i,
-                      3 + 4 * i,
-                      0 + 4 * i,
-                      2 + 4 * i } };
+    QuadIndex y = { { static_cast<uint16>(0 + 4 * i),
+                      static_cast<uint16>(1 + 4 * i),
+                      static_cast<uint16>(2 + 4 * i),
+                      static_cast<uint16>(3 + 4 * i),
+                      static_cast<uint16>(0 + 4 * i),
+                      static_cast<uint16>(2 + 4 * i) } };
     quad_index_list[i] = y;
   }
